move errors into the error file
